4446 ftp - test - refactor - extract BaseFtpActionIntTest and refactor tests
